+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### :syringe: Fixed
+
+- Action fails when assignees are configured
+- Action fails when labels are configured
+
 ## [1.3.0] - 2020-07-28
 
 ### :zap: Added